pyproject.toml
[project]
name = "planindicativo-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "django~=5.2",
    "django-unfold>=0.78.1",
]
